+++ smartcontract/token.hpp
@@ -29,7 +29,7 @@ public:
 			uint64_t id;
 			string name;
 			account_name issuer;
-			uint64_t issuer2;
+			uint32_t issuer2;
 			uint32_t limit;
 			uint8_t when;
 			uint8_t disposal;
@@ -56,7 +56,7 @@ public:
 	 */
 	[[eosio::action]]
 	void tknew(const account_name sender, const string& name, const account_name issuer, 
-			   const uint64_t issuer2, const uint32_t limit, const uint8_t when, 
+			   const uint32_t issuer2, const uint32_t limit, const uint8_t when, 
 			   const uint8_t disposal,const uint8_t type, const uint64_t taskid, 
 			   const uint64_t tokenid, const uint32_t reftoken, const string& term, 
 			   const uint8_t rcalctype, const uint32_t nofdevtoken );
@@ -67,7 +67,7 @@ public:
 	[[eosio::action]]
 	void tkupdate(const account_name sender, const uint64_t id,
 			   const string& name, const account_name issuer, 
-			   const uint64_t issuer2, const uint32_t limit, const uint8_t when, 
+			   const uint32_t issuer2, const uint32_t limit, const uint8_t when, 
 			   const uint8_t disposal,const uint8_t type, const uint64_t taskid, 
 			   const uint64_t tokenid, const uint32_t reftoken, const string& term, 
 			   const uint8_t rcalctype, const uint32_t nofdevtoken );
